[ports-mingw] gdk-pixbuf: properly export the symbols
Before we were mandatorily removing some of the gdip modules. This is
bad since they are actually pulled in when using the included loaders
and anyway they are only used as fallback modules for when the ones
using the free software libraries are not available.

Source: msys2/MINGW-packages@7bf3b0042a61aca5b9a239f8923b3485e81e50c6