[ports-mingw] creduce: Added (WIP)
This is a weird package. The same PKGBUILD is
used to make both the mingw-w64 C++ program and
also the MSYS2 perl packages. There are two
perl packages for each mingw-w64 package because
they hardcode /mingw{32,64} into them. This
could be changed via an env. var set by creduce
which is itself a perl script that lives in
the /mingw{32,64}/bin folder.

Apparently, differences in makepkg*.conf makes
this scheme not work, so we need to either fix
that or split it into two PKGBUILDs.

The problem with that is that this will never
build as an MSYS2 PKGBUILD as we don't have
clang there, so the build will fail and we'll
not be able to install any of the perl modules.

Source: msys2/MINGW-packages@d7e338759dab3aa63c37c88764ec3b396c4837e4