[ports-mingw] qt5-static: Consider junction points as symlinks
This reverts a reversion of a commit that added this
originally. The reasons for the reversion seem weird
and only hand-wavingly explained - we discussed this
on qt's private dev IRC channel and decided this was
a good idea - roughly.

This is the first part of a fix for:
https://github.com/msys2/msys2.github.io/issues/4

Source: msys2/MINGW-packages@46de770a40e2f844faa215ee19d046774dcfc246