[ports] pacman: Sanity check the final DB URL
When using the *internal* downloader the returned `final_db_url` is
appended by the postfix `.sig` without checking if the `final_db_url` is
indeed sane.

Possible redirects by the server *could* have altered it and therefore
it should be checked before being processed any further. For example,
Bintray redirects all of its URLs, but it is not a dynamic redirector,
as is visible by the URL: it does not end in `.sig` at all, but rather
is a long hexadecimal string.

The `pacman` ML already contains a patch for this kind of behavior. This
commit will apply that patch before building `pacman` for *MSYS2*.

Signed-off-by: nalla <nalla@hamal.uberspace.de>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Source: msys2/MSYS2-packages@0b2c66dc0ac9fab7ade55bf301cbadfbb78d18ee