[ports] bash-completion: avoid unnecessary tilde expansion
When connected to an Active Directory (a variant of LDAP), tab completion
becomes very slow because the tilde is expanded all the time, even if
completing paths that do not contain a tilde at all. This developer
observed delays of several seconds at a time.

This is completely unnecessary, so let's do not do that.

The patch was provided by Ove Risberg with a bug report to the original
bash-completion project:

	https://bugs.launchpad.net/ubuntu/+source/bash-completion/+bug/1390061

Since it fixes code introduced by another add-on patch, that add-on patch
was modified instead of adding yet another patch.

This fixes https://github.com/git-for-windows/git/issues/377

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Source: msys2/MSYS2-packages@ec3da11425d532992f1a6e2071b1aca246a6568f