[ports] winpty-git: Restore path mangling of all command-line arguments
In 35510397c6e92845dbad8cdb215d3bef67dc2529, patch 0002-Add-support-for-conversion-of-all-argv-path-args.patch
was removed because it was thought to be superseded by an upstream commit[1], but that was actually only one
part of the patch. This commit restores the patch minus the upstreamed part. Fixes #360.

[1] https://github.com/rprichard/winpty/commit/2ecc596f8b191b8b8cd90dfffba162985d3ed0e3

Source: msys2/MSYS2-packages@e270bb5d3c87cf45eabf4bcad5cf86f05e661164