[ports] Enabeling GSSAPI/Kerberos authentication for git-svn
Curl is used as a wrapper for the serf library?! I'm not 100% sure about this but without Kerberos support enabled for curl, compile switch --with-gssapi, the autentication is not working when accessing a kerberised Subversion server via "git svn ...". To test curl in a standalone mode use the --negotiate parameter. To pass a Kerberos authenticating proxy pass the argument --proxy-negotiate. On a Windows computer which is member of a domain a Kerberos tgt is automatically available after login. For none domain Windows coputers this is NOT working!

Source: msys2/MSYS2-packages@a4e2d7f0c82585357f792cbe381a5a4d77f34761