[ports] Enabeling GSSAPI/Kerberos authentication for git-svn
Serf is the library used for HTTPs communication of a client with a Subversion server. This library is also used by git-svn by the curl wraper?! That serf can communicate/authenticate with/to a kerberised Subversion server GSSAPI must be enabled for scons. In the change some more dependencies were added that the compilation is running.

Source: msys2/MSYS2-packages@ef0374ad98fc389136412e900b89402269e28791