[ports] gyp-git: Replacement for gyp-svn
I think msys-python multiprocessing is broken though
as it calls mingw-w64-python instead of msys-python.
This is not a in for this package, but it might be
wise to prevent gyp from using multiprocessing until
it is fixed.

Source: msys2/MSYS2-packages@b7ff8b795a95bfe14cf54dd8f5cdac85b434a090