[ports] mpfr: Fix versioning
mpfr package had invalid version 3.1.4.-3 when it should have been
3.1.4-3. This was caused because an empty _patchlevel was set appended
to the package version separated with a dot.

This commit changes the package versioning so that the _patchlevel is
appended to the package version only if it isn't empty.

Similar commit 0c86b53 was done to mpfr package in MINGW-packages.

Source: msys2/MSYS2-packages@2f187729af77224bb80064a68b1e58f16631973e