[ports-mingw] Update mpv to 0.23.0 (and update mpv's dependencies) (#2060)
* speex: Update to 1.2.0

The WinMM patch has been upstreamed.

* wavpack: Update to 5.0.0

* libwebp: Update to 0.5.2

* harfbuzz: Update to 1.3.4

* freetype: Update to 2.7.1

* libass: Update to 0.13.6

* ffmpeg: Update to 3.2.2

This adds --enable-libwebp as an argument to the configure script. It
was probably always supposed to be there because libwebp is a
dependency, but ffmpeg doesn't seem to link to it unless the argument is
also specified.

* libcdio: Update to 0.94

* mpv: Update to 0.23.0

Packaging changes:
- enca and libguess are removed in favour of uchardet
- CDDA support is re-enabled, since it appears to work now
- The licence field is changed back to "GPL." This should probably have
  been done in 0a1af2b, when the NNEDI3 prescaler was removed.
- --disable-egl-angle-lib is used to prevent linking directly to libEGL,
  because this doesn't work on i686. Instead, libEGL will be loaded at
  runtime.

Source: msys2/MINGW-packages@0141a98caf610ae3b0c23a521b70fd63d8ddef3b