[ports-mingw] python3: Make multiprocessing.Queue work; Add some tests. Fixes #837 (#2619)
With non-mingw builds POSIX_SEMAPHORES_NOT_ENABLED is not enabled and
sem_unlink is a no-op. Change it to always be provided on Windows instead.

This adds a minimal test suite which makes it easier to catch regressions
in the future for common MSYS2 related bugs.

Source: msys2/MINGW-packages@3f6c286d6a1232eb9581b3a831169fd4ea880010