[ports] coreutils: fix libstdbuf.so.exe problem once and for all
The `stdbuf.exe` program links to a `libstdbuf.so` file, so if that does
not exist, there is a problem.

In d9773f0a (coreutils: Update to 8.24, 2015-08-28), a sed call was
introduced to ensure that `libstdbuf.so` is built and installed under
that name, without `.exe` suffix.

As reported in https://github.com/Alexpux/MSYS2-packages/issues/541,
however, that fix no longer works, the `libstdbuf.so.exe` file is in
coreutils 8.26-1.

Let's just rename the `libstdbuf.so.exe` file after it has been
installed into the `pkgdir` tree.

Note: we still have to muck around with `Makefile.in` so that the build
does not break with

	make[2]: *** No rule to make target 'src/libstdbuf.so', needed
	by 'all-am'.  Stop.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Source: msys2/MSYS2-packages@7e7afe092234b380a15dd735ce042f804b6fd14d