[ports-mingw] CMake 3.10.1 (#3262)
* cmake: Upgrade CMake package to 3.10.1

0004-Implement-Qt5-static-plugin-support.patch does not apply cleanly.

* cmake: Rebase the patch series onto 3.10.1

The patch series was applied using `git am` to a CMake v3.9.6 git
checkout, and then those commits were rebased onto CMake v3.10.1.

`git format-patch` was used to extract the patches again, as was
presumably originally done. In this case, I kept the existing numbers to
keep the patch readable, and avoid confusing file renames.

Only 0004-Implement-Qt5-static-plugin-support.patch needed manual
changes.

* cmake: CMake now unconditionally enables server mode

* cmake: Remove dead unit-test munging

This code appears to have been updating the paths in a CMake Unit Test
to match the live paths in MSYS2.

However, it was disabled when added (see #2633), and doesn't make sense
since the test is only of the parsing of the output format, not the
contents of the output.

Source: msys2/MINGW-packages@1ad389907cc2203002a04fea750bba7fc02430b6