[ports-mingw] glib2: Update to 2.54.3; Switch to Python 3 for the build tools (#3307)
Python 3 is the default on Linux for most distros now, so it's a good idea
to use the same tools as everyone else.
I've built a few packages depending on glib and everything seems to work with py3.

Source: msys2/MINGW-packages@9c4bfaceed3a621bb83062bce50b431eb8bf3c33