[ports-mingw] Teach cURL to find the certificate bundle also when validating the certificates of proxies (#3592)
* curl: ensure that the correct ca-bundle.crt is used for HTTPS proxies

There are actually two separate settings for the ca-bundle.crt file: one
to validate HTTPS servers, and one to validate HTTPS proxies. We forgot
the proxies.

This is a port of the fix for
https://github.com/git-for-windows/git/issues/1493

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

* curl: increase pkgrel

We just fixed the long-standing issue where cURL would not find the
certificate bundle when trying to validate HTTPS *proxies*'
certificates.

Let's increase the pkgrel in preparation for a new build.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Source: msys2/MINGW-packages@5a77159a52dbfac078cbdd9410b8f49be1598892