[ports-mingw] meld3: fix "has no attribute Win32FontMap" error on start. Fixes #3610 (#3611)
I think this type was supposed to be private in pango and not exported in the gir.
But it was anyway with the autotools build and now no longer is with the meson one.

The code in question was trying to set the win32 pango backend which is also
possible through an env var, so just do that instead and make sure we do it
before importing pango.

Source: msys2/MINGW-packages@7b16f74d1e62f13d784f31cb6b0cfea9f7d469e2