[ports-mingw] cmake: Update to 3.11.1 (#3545) (#3628)
The patch series was recreated by applying it to a CMake git checkout at
v3.10.1, and then rebasing it onto v3.11.0, followed by patch resync
with CMake 3.11.1

Patch 0004-Implement-Qt5-static-plugin-support.patch had to be
extensively redone, to support a refactoring in the Qt5 auto-gen support
in CMake 3.11.0. The rest of the patches are changed only due to
artifacts of this rebasing.

Source: msys2/MINGW-packages@d4098577816cfab1353e3b85e8b929724734b76b