[ports-mingw] Upgrade node.js to v8.11.1 (#3646)
* nodejs: re-generate patchset

Git's `format-patch` now uses a narrower diffstat, longer abbreviated
hashes in the index lines, and a different Git version, of course.

Besides, the patches started out as an 11-strong patch series (and is
now 13-strong, which is reflected in the [PATCH k/13] prefix of the
Subject: lines), and the 12th patch even lacked attribution and a commit
message.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

* nodejs: adjust patches for v6.14.1

The code base changed enough that some patches are obsolete, and others
need to be applied to a different file, yet others simply do not apply
cleanly.

Note: this does not yet allow building v6.14.1, but *only* adjusts the
existing patches as required, for ease of reviewing.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

* nodejs: do not link to the shared (stale) libuv

Node.js v6.14.1 relies on libuv to be up to date. In particular, it
requires the uv_os_getppid() function to be available.

The version of libuv shipped in MSYS2 is too old and does not provide
that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

* nodejs: update to v6.14.1

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

* nodejs: prepare for packaging v6.14.1

Long path problems...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

* nodejs: adjust patches for v8.11.1

Sadly, v6.14.1 would not be able to package because of too-long paths
inside npm's nested node_modules/ directories.

Let's hope that v8.11.1 works better.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

* nodejs: prepare for v8.11.1

It needs quite a bit more patching.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

* mingw-w64-nodejs: upgrade to v8.11.1

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Source: msys2/MINGW-packages@87a0b47b97cdf2192f3fe9c32d574fd521cb8cc8