[ports-mingw] meson: fix python module in case the official Python is installed (#4064)
meson uses the py.exe to figure out the path of the default python2/3
but msys2 python doesn't install it. This leads to it finding the
system Python and getting the wrong path.

This fixes the gobject-introspection build with meson (not the default yet..)

Source: msys2/MINGW-packages@8512d47455ffd82f732d621988f2df69956e2d94