[ports-mingw] Revisions to the mingw-w64-python package template. (#4066)
Add export SETUPTOOLS_SCM_PRETEND_VERSION=$pkgver to prepare for setuptools-scm if necessary.  That's required for release versions (not a git repository)
Clearly denote a section in the python2 and python3 install sections for .exe's and other scripts that are deployed in the /mingw*/bin directories  Also note what needs to removed for packages that do not do this.
For python2 scripts and .exe's, add code to rename scripts and executables to avoid a name conflict with python3 packages.

Source: msys2/MINGW-packages@9e95e78eb9de5bb86b4aaefc36b91b3e8e6f95ab