[ports-mingw] have the mingw64 pkg-config installed with cmake
when running under mingw64.exe/mingw32.exe the pathes from the environment are
transformed into Windows pathes. As a result PKG_CONFIG_PATH because a Windows
path. When passed to the non-mingw pkg-config the path becomes invalid and is
not used properly.

To solve this the mingw pkg-config must be installed. It works with Windows
pathes as expected.

As a background to this issue, this was originally reported to cmake:
https://gitlab.kitware.com/cmake/cmake/issues/18980

Source: msys2/MINGW-packages@c4d770d7642d576bd31b61d04416847d15f9f9c5