[ports-mingw] ruby - 2.6.3 - added provides.
Added provides for the default gems and the bundled gems that are included as well as a script for updating them.  The idea is that as more gems get added, we could track dependencies or gems that are already included.  This is similar to what is being done with perl.  If ruby drops a gem from the bundle or default, we want to know in case a dependency is broken.  Some distributions actually package bundled gems separately.

Source: msys2/MINGW-packages@dc27816c5835037ae7d79e7a85efa7d8ec7eb8a1