[ports] Upgrade the MSYS2 runtime to v3.1.4 (#1953)
* msys2-runtime: add a shell script to generate the patches from msys2/Cygwin

The idea is to commit a change incrementing the `pkgver`, then running
this script, and then amending the commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

* msys2-runtime: work around Cygwin's source code containing symlinks

It is a bit unfortunate that MSYS2 substitutes symlinks by deep copies,
by default. Git, for example, claims that it cannot do symlinks, but
Cygwin now needs this. Let's make do with that default.

The strategy we use is simple: we override the auto-detected
`core.symlinks` setting. That will result in a dirty worktree, of
course: deep copies are not symlinks, and `git status` will detect it.
But that does not matter, as the build will now work ("deep" copies of
those header files will work just as well as real symlinks would have
done).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

* msys2-runtime: upgrade to v3.1.4

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Source: msys2/MSYS2-packages@5dea2a93518b9abd34553ca2600d20847ca4c5e1