[ports-mingw] gobject-introspection: fix g_ir_scanner var in .pc file
We create an .exe launcher for it so the variable in the .pc file
no longer matched.

This is a problem because meson 0.55.0 has started to use the variable
to look up g-ir-scanner and fails.

Fixes #6677

Source: msys2/MINGW-packages@56c0932ff798b5b5d5afd08dedf4d3aba1b58af8