[ports-mingw] Add gtk-update-icon-cache package
This just contains the gtk-update-icon-cache binary which doesn't
depend on gtk itself and the pacman hooks for updating the cache.

This allows other packages to depend on it for getting the hooks without
having to depend on gtk3.

This mirrors what Arch Linux does.

Source: msys2/MINGW-packages@84c96dc3e201a4026bb8f97348c00dde313df686