[ports-mingw] python: copy all python launchers into the venv
Turns out Python uses the basename of the calling executable to
decide which launcher in the venv it uses to call pip with.
This means we have to copy all launchers we ship.

Use the non versioned launchers as src so the rewrite logic in venv
copies the venvlauncher variants.

Fixes #7227

Source: msys2/MINGW-packages@7ab734e8fd908c54014cd7a9193cff5a6672673e