[ports-mingw] python: cygpty hack improvements
In case of cygpty we want to fake a tty and at the same time make
Python behave as if it was attached to a console.

This improves two thing in that regard:

* Update the is_cygpty patch to cover more isatty() checks (likely
  they were added since the patch was created)
* Default the stdio encoding to the console output encoding by default.
  This is only in effect if stdout isn't redirected (then is uses the ANSI code page)
  and no real console is used (then it forces utf-8). Python never considers the
  case were isatty() and no console, so we have to add this.

This makes printing unicode work again in MSYS2 bash.

Source: msys2/MINGW-packages@55bb9c2d1df9bf8abeff8fbadc3b8f3b6522b30d