[ports-mingw] CMake: Executables need to generate import libraries
It is not uncommon for executables to allow the use of plugins, and
plugins need to access symbols exported from the executable.  On windows,
things require explicit linking, therefore this import library is
necesary to be generated in order to be able to compile the plugins.

At one point in time, MinGW had some linking issues, and it appears
this patch was put in place to disable generation of import libraries
for executables due to lack of understanding of the consequences.

CMake's author's agree with this assessment as they did not accept
this merge request:
https://gitlab.kitware.com/cmake/cmake/-/merge_requests/5194

Source: msys2/MINGW-packages@76446f4dfeffbfc47ffb5b428c2a7b288b49a44b