[ports-mingw] gh: upgrade to v1.9.0
New goodies! In particular `gh run view`...

Also: With this version, we no longer need that `Makefile` patch, but we
do need to set `GOOS` so that the file extension `.exe` is used, and
then ask for `bin/gh.exe` to be built, not `bin/gh`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Source: msys2/MINGW-packages@6b4901110a1975505dab8cbd5336675c691d27d3