[ports-mingw] curl: use plain copies of mingw-w64-pathtools's sources
This patch changes the files so that the pathtools source code from the
`mingw-w64-pathtools` directory is used.

Sadly, we cannot use relative paths (which is forbidden by `PKGBUILD`)
nor symbolic links (because of MSYS2's inadequate support for symbolic
links). Therefore we do the next best thing: we copy the code and add a
check to ensure that the code hasn't gone stale.

Note: This is not _quite_ an idempotent change, as the
`mingw-w64-pathtools` folder offers the `get_relocated_path()`, the
`single_path_relocation()` and the `pathlist_relocation()` function,
which cURL does not need. However, in the interest of avoiding
duplication, this is the right thing to do, and the other functions'
implementations are identical to what was in
`0001-Make-cURL-relocatable.patch`.

Also note: an earlier attempt wanted to hard-code a certain
MINGW-packages commit from which to download the pathtools via https://
from GitHub, i.e. despite cloning the repository and having the source
code locally already, the `PKGBUILD` would insist on loading a
potentially different revision from elsewhere. However, this would cause
problems in the future when contributors want to open a PR to update the
pathtools _and test before pushing_ (which is the good practice,
though). Therefore we use copied files instead.

Co-authored-by: Biswapriyo Nath <nathbappai@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Source: msys2/MINGW-packages@31374aacd434170b03a41bf4bb45fa9d88d2d348