[ports-mingw] gnupg: remove unused patch
We are about to consolidate all locations where the
`mingw-w64-pathtools` are used, and `05-mingw-relocation.patch` _looks_
like it teaches gnupg to do that. But the patch is not even applied, and
the gnupg code is not taught to use the pathtools functions, either.

So let's just drop the patch unceremoniously.

Source: msys2/MINGW-packages@93bb57f42625a099365091aaa80b28ba5c6abe7d