[ports-mingw] extra-cmake-modules: pass --force-local to GNU tar
It looks for GNU tar, and if found uses it instead of cmake tar
functionality, in order to produce a "reproducible" tar file.
Unfortunately, GNU tar assumes a filename with a colon in it is a remote
file (host:path), not a Windows drive letter (X:\path).  Pass the
--force-local option to tell it not to make that assumption.

Source: msys2/MINGW-packages@b2edb3830c357c35c953bc60744a6faf7a7a3dad