[ports-mingw] reinstall splitpkg template in PKGBUILD files
On packages which contained the template contents but were missing the
template start comment that tells makepkg-template that it is in fact
the template.  This will allow makepkg-template to be able to update the
template in these files if it ever should change.

Source: msys2/MINGW-packages@7a35c4cc17c8a0242d9490d25082de32261c8422