[ports-mingw] xpm-nox: fix #8548 XPM writes invalid X Pixmap with path as C identifier
The Windows path characters : and \ were not filtered out when writing
the C identifier that is part of the XPM image.
Besides that writing the complete path inside an image can
lead to a C identifier which is too long.

So let's just remove the path and only leave the filename.

Source: msys2/MINGW-packages@7505519157a3296e46e2261fc198d00606637523