[ports] Add an autotools meta package
The goal of this package is to depend on everything that is (usually)
needed for running autogen/autoreconf if needed,
configure and make, but not a compiler.

If packages would makedepend on this then we could make base-devel smaller.

Source: msys2/MSYS2-packages@ea223c5f2c071baff4510b1ffa1626d1615faf33