[ports] add binutils to base-devel
makepkg requires a strip.exe if strip is in options before it can install
build dependencies. So we have no choise but to assume that it is installed.

An alternative would be to patch pacman and move the check to the actual
strip function tidy_strip()

Source: msys2/MSYS2-packages@e6cf475a0ffbcaf0032340a22598124e3bee0023