[ports] libtool: override the soname_spec for cygwin as well
So that if we target a cygwin triplet we still get libraries with a "msys-" prefix.

This allows us to get rid of some msysize patches if we run autoreconf and
pass a cygwin triplet instead of a msys one.

Source: msys2/MSYS2-packages@c54b9e9fc276f3faa498f39ee3b45122a0ff64ae