[ports-mingw] gettext: do include the libintl_*wprintf() functions
They are declared, after all.

Fun fact: in v0.19.8.1, those functions were included in the library,
but not declared in the header. In v0.21 the were _not_ included in the
library, but they _were_ declared in the header.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Source: msys2/MINGW-packages@028411f0317ee8678acc0cb5a51c593f3293bc9e