[ports] glib: Update to 2.70.5
Drop the gconvert patch as it was rejected upstream and it's unclear
what exactly it fixed for us.

Drop the gmodule patch as it's unclear if it's still needed, and module
loading seems to work for me.

Add a small patch replacing the cygwin DLL prefix. There are more cases
in the tests where MODULE_FILENAME_PREFIX and G_MODULE_SUFFIX are wrong,
or not set for cygwin, but that's for another time.

Source: msys2/MSYS2-packages@7b33d268cc07ea176f5db34f5684cdca6ed0c6b2