[ports] base: add tar
Windows nowadays has a tar.exe in System32 which is first in PATH
with a clean MSYS2 install.

This will both confuse users as well as scripts because it is
a non-cygwin bsdtar.

To avoid surprises add tar to base.

Source: msys2/MSYS2-packages@44fe93f472f84a8a3cd938eb175adc81fdbc5120