[ports-mingw] emacs: some hacky patches to make it work on ucrt64
for both stdout and stderr fclose fail during atexit as if they
were already closed. the function in theory has a workaround for that
in that it will continue if errno is EBADF, but fclose doesn't set that
on Windows.

This shows at least what is missing/broken if upstream wants to have a look.

Source: msys2/MINGW-packages@01be90ce74e3b69e7b03ab886d5f4ac59da0f646