[ports-mingw] CI: upload a process minidump in case something crashes
In case something crashes in CI this will show us some
information on which program crashed and why.

We register procdump as a JIT debugger which dumps the process
memory into a new file if triggered. To enable the JIT debugger
in the context of cygwin we need to set MSYS=winjitdebug.

The resulting dump file can for example be analyzed with
minidump-stackwalk or minidump_stackwalk

Source: msys2/MINGW-packages@6ff1a1f8a631ab9b243a827e77619dc2f33cb992