[ports-mingw] Add the `mingw-w64-7zip` package
MSYS2 does offer the `p7zip` package. But it is an MSYS package (and
therefore the executable is slower than it needs to be), and besides,
the p7zip project seems to be abandoned since 2016.

In the meantime, the 7-Zip project itself added support for building the
command-line tool on Linux, and we can use that very same setup to build
a MINGW version, too. So let's do that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

Source: msys2/MINGW-packages@a023812d5a851da1f4521044efce37a54582e10d