[ports-mingw] rust: set _bootstrapping=no for clangarm64.
Either this, or we'll need to continue to inject new stage0 packages
during rust updates.  I left the patch and code to do that in place for
now in case that turns out to be necessary or desirable.

Source: msys2/MINGW-packages@4dffb82fbe6b760bbbe945ab1f2ebc58cbc3e6df