[ports] tcl: Update to 8.6.12
Partly sync with cygwin (the patch) and Arch (the packaging)

It's a bit mysterious where the cygwin patch is coming from,
it's in the src package, but not the git repo. And it likely
comes from some github repo, but I can't find it.

Source: msys2/MSYS2-packages@c9c32dcb31609662c20f0cf4a2e97408671f8682