[ports] Remove perl-Crypt-SSLeay
No users left. We'd need to patch if for openssl v3 support, and
I'm not sure it's worth it.

If it turns our we still need it in the future we can bring it back
and backport some opensslv3 patches from
https://src.fedoraproject.org/rpms/perl-Crypt-SSLeay/tree/rawhide

Source: msys2/MSYS2-packages@dc01ff654b65229a40e70ec87fd2da4e1bb222c4