[ports-mingw] Add back CMake configuration files
Boost provides CMake configuration files. These files are the ultimate truth
about how Boost libraries are to be compiled. It's better to rely on them, then
on CMake provided FindBoost.cmake module. The latter can become stale, or it can
contain bugs, which will make users dependent on the ability of CMake authors to
fix it. Boost provided configuration files are always up-to-date and tested with
the version we are installing.

This PR also adds patch to the list of build dependencies.

Source: msys2/MINGW-packages@2beb521be689dd2b637ac91c23a1d9562126ba52