[ports] readline: add etc/inputrc
This is just a copy of the current inputrc from the "filesystem" package
which currently gets installed as ~/.inputrc on the first start.

The goal is to remove ~/.inputrc, so that on the first msys2 start the config
is already there and a shell restart is no longer needed.

readline is statically linked into bash, so there is no readline dependency,
so we assume that libreadline is installed with "base" in some way.

Source: msys2/MSYS2-packages@fd25584fe30bc4fc9f429e9917593f2fbd6b9b03