[ports] filesystem: remove ~/.inputrc and remove the restart warning
We now provide ~/.inputrc as /etc/inputrc since fd25584fe30bc4fc

This means that on the first start we will execute /profile, which
writes various  bash related files to $HOME, which bash will execute
after /profile is finished.

So with ~/.inputrc out of the way the environment on the first start
should be the same as the second one, so we can remove the restart
warning at the beginning.

The question remains if we should still call "clear" at the end, hiding
the setup process. I decided for removing it and enclosing the setup with
some colored status message. This hopefully is not too confusing for
users seeing the intial process and still allows an easy way to see the logs
if something goes wrong.

Source: msys2/MSYS2-packages@f7d14586f177a0cce120a064692d11de4a42f82f