[ports-mingw] Add clang32 to mingw_arch defaults
And while at it also add explicit mingw_arch to all PKGBUILD
files that don't have it already.

So that we get clang32 by default if mingw_arch is missed.

Not clangarm64 because we can't test in the PR CI, but maybe
we should add it anyway, not sure.

Source: msys2/MINGW-packages@0b20739e87f353b2e871bcb2f51f52e6633a251b