[ports-mingw] Added SDL2 is a build dependency in PKGBUILD.
SDL2 is only a runtime dependency if the sdl display host is enabled, but the configure check requires it to pass for some reason. This should be fixed in a future revision, but for now we'll just require it for building.

Source: msys2/MINGW-packages@4e40e665689525da1fe8051c6d5bc27576b13c6b