[ports-mingw] gcc: Remove libssp-0.dll
There are no packages in MSYS2 that use it for 7 months, and
we don't ship an import library for it since then as well,
so remove it.

This also makes it possible to build gcc with stack protector enabled.

Source: msys2/MINGW-packages@dfa6dd8c269cb4899e57f7ae033ab54e5a7b414f