[ports] mingw-cross-binutils: remove msysize patch
build for cygwin instead

autoreconf is broken with this version so we can't add the patches
for the DLL naming, so the bfd libdep plugin has cygwin naming.
Luckily binutils loads all DLLs in that directory so it shouldn't
matter.

We'll try to autoreconf in future versions.

Source: msys2/MSYS2-packages@13909e7100bbdca9770f3b60a506a586e7e6a192