[ports-mingw] libpaper: don't write a /etc/papersize  by default
this was copied from Arch, but the parsing code doesn't understand
empty lines and comment lines, so I'm not sure how that is supposed to work.
This led to the default papersize being the comment line which makes
the lookup fail, and in combination with xpdf not checking any return values
results in xpdf crashing.

Source: msys2/MINGW-packages@afd3ceb085e7e68f65bc02be2a017285b1a4f7a4