[ports-mingw] python-setuptools: add support for py_limited_api
This makes distutils link against python3 instead of python3.x when
py_limited_api is set on the extension. With MSVC this is solved via
pragma, which doesn't have an alternative with our toolchains.

This depends on libpython3.dll being installed, so
https://github.com/msys2-contrib/cpython-mingw/pull/148

Source: msys2/MINGW-packages@cba50507490805827ce9bec641354669173ae2f0