[ports-mingw] qt5-connectivity: Fix building in 32 bit environment
Once upon a time, the PFN_BLUETOOTH_ENUM_ATTRIBUTES_CALLBACK was defined
in mingw-w64 incorrectly. It was fixed in this commit[1].

But, Qt5 worked around that issue. Now that workaround is not valid.
This change removes that workaround.

[1]: https://github.com/mingw-w64/mingw-w64/commit/427ab730f00c2dc45508f0e27af7f7a67cf9f3ab

Source: msys2/MINGW-packages@ec6dda0bd483fb45d37f2438723913eee7e9b803